requirements.txt
Pillow
pywin32